src/bot.py
# Logic for the M8N-Dating Bot
import discord

class DatingBot:
    def __init__(self):
        self.client = discord.Client(intents=discord.Intents.default())

    def run(self):
        token = ""  # Load token from config.settings.py
        self.client.run(token)