requirements.txt
# Basic dependencies for the M8N-Dating Bot
discord.py==2.1.0
python-dotenv==1.0.0